% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegression.R
\name{coefficients}
\alias{coefficients}
\title{calculate coefficients of multiple linear regression}
\usage{
coefficients(X, Y)
}
\arguments{
\item{X}{A matrix representing the design matrix}

\item{Y}{A numeric vector of values as response variable}
}
\value{
A matrix of regression coefficients(includes estimates,standard error, t statistic and p value)
}
\description{
calculate coefficients of multiple linear regression
}
\examples{
Y <- c(5,7,10,5)
X1 <- c(1,2,3,4)
X2 <- c(2,3,4,6)
X <- cbind(1,X1,X2)
coefficients(X,Y)
}
