% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegression.R
\name{fitted_values}
\alias{fitted_values}
\title{calculate the fitted values of simple linear regression}
\usage{
fitted_values(X, Y)
}
\arguments{
\item{X}{A numeric vector of values as predicted variable}

\item{Y}{A numeric vector of values as response variable}
}
\value{
fitted values of simple linear regression
}
\description{
calculate the fitted values of simple linear regression
}
\examples{
X <- c(1,2,3,4)
Y <- c(2,3,4,5)
fitted_values(X,Y)
}
