% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegression.R
\name{Residuals}
\alias{Residuals}
\title{calculate the residuals of linear regression}
\usage{
Residuals(X, Y)
}
\arguments{
\item{X}{A matrix representing the design matrix}

\item{Y}{A numeric vector of values as response variable}
}
\value{
A vector representing residuals of linear regression
}
\description{
calculate the residuals of linear regression
}
\examples{
Y <- c(4,6,9,13)
X1 <- c(2,3,5,6)
X2 <- c(1,4,3,8)
X <- cbind(1,X1,X2)
Residuals(X,Y)
}
